requirements.txt
# UI
streamlit>=1.32.0
python-dotenv>=1.0.0

streamlit-aggrid==0.3.5

# LangChain stack (versions must match)
langchain==0.3.27
langchain-community==0.3.27
langchain-openai==0.3.29
openai>=1.30.0

# Vector backends
faiss-cpu>=1.7.4
# Install Chroma only off-Linux (i.e., for your local dev) to avoid sqlite on Cloud
chromadb==0.5.5; platform_system != "Linux"

# Retrieval helpers
rank-bm25>=0.2.2
tiktoken>=0.7.0

# Ingestion
pdfplumber>=0.11.0
beautifulsoup4>=4.12.3
lxml>=5.2.2
requests>=2.32.3

# Google Sheets logging
gspread>=6.0.0
google-auth>=2.30.0
